lambda/set_webhook/requirements.txt
httpx[http2]>=0.27.0
python-telegram-bot~=21.2
boto3~=1.34.116
